# docs/swagger.yaml
basePath: /api
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anheyu.com
    name: 安知鱼
    url: https://github.com/anzhiyu-c/anheyu-app
  description: Anheyu App 应用接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Anheyu App API
  version: "1.0"
paths:
  /albums:
    get:
      description: 获取相册图片列表，支持分页、标签筛选、时间筛选和排序
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 标签筛选
        in: query
        name: tag
        type: string
      - description: 开始时间 (2006/01/02 15:04:05)
        in: query
        name: createdAt[0]
        type: string
      - description: 结束时间 (2006/01/02 15:04:05)
        in: query
        name: createdAt[1]
        type: string
      - default: display_order_asc
        description: 排序方式
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取相册图片列表
      tags:
      - 相册管理
    post:
      consumes:
      - application/json
      description: 新增图片到相册
      parameters:
      - description: 图片信息
        in: body
        name: body
        required: true
        schema:
          properties:
            bigImageUrl:
              type: string
            bigParam:
              type: string
            displayOrder:
              type: integer
            downloadUrl:
              type: string
            fileHash:
              type: string
            fileSize:
              type: integer
            format:
              type: string
            height:
              type: integer
            imageUrl:
              type: string
            tags:
              items:
                type: string
              type: array
            thumbParam:
              type: string
            width:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 添加失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 新增相册图片
      tags:
      - 相册管理
  /albums/{id}:
    delete:
      description: 根据ID删除相册中的图片
      parameters:
      - description: 图片ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: ID非法
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除相册图片
      tags:
      - 相册管理
    put:
      consumes:
      - application/json
      description: 更新相册中图片的信息
      parameters:
      - description: 图片ID
        in: path
        name: id
        required: true
        type: integer
      - description: 图片信息
        in: body
        name: body
        required: true
        schema:
          properties:
            bigImageUrl:
              type: string
            bigParam:
              type: string
            displayOrder:
              type: integer
            downloadUrl:
              type: string
            imageUrl:
              type: string
            tags:
              items:
                type: string
              type: array
            thumbParam:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误或ID非法
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新相册图片
      tags:
      - 相册管理
...
securityDefinitions:
  BearerAuth:
    description: '在请求头中添加 Bearer Token，格式为: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"